third_party/WebKit/LayoutTests/css3/calc/getComputedStyle-border-radius-expected.txt

Tests calling getPropertyValue on computed border radii involving CSS calc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-top-left-radius') is getComputedStyle(document.getElementById('control', null)).getPropertyValue('border-top-left-radius')
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-top-right-radius') is getComputedStyle(document.getElementById('control', null)).getPropertyValue('border-top-right-radius')
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-bottom-left-radius') is getComputedStyle(document.getElementById('control', null)).getPropertyValue('border-bottom-left-radius')
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-bottom-right-radius') is getComputedStyle(document.getElementById('control', null)).getPropertyValue('border-bottom-right-radius')
PASS getComputedStyle(document.getElementById('calc'), null).getPropertyValue('border-radius') is getComputedStyle(document.getElementById('control', null)).getPropertyValue('border-radius')
PASS successfullyParsed is true

TEST COMPLETE
